removed feature that not working and that i don't plan to continue the support
