+++ bot/src/background_task/server_image/calculate_user_color.rs
@@ -8,7 +8,6 @@ use crate::database::prelude::UserColor;
 use crate::database::user_color::{ActiveModel, Column, Model};
 use crate::event_handler::{add_user_data_to_db, BotData};
 use crate::get_url;
-use crate::new_member::change_to_x128_url;
 use base64::engine::general_purpose;
 use base64::Engine;
 use futures::stream::FuturesUnordered;
@@ -28,6 +27,25 @@ use serenity::nonmax::NonMaxU16;
 use tokio::sync::RwLock;
 use tokio::time::sleep;
 use tracing::{debug, error, trace};
+
+pub fn change_to_x128_url(url: String) -> String {
+	debug!("Changing URL size to 128x128: {}", url);
+
+	let mut url = url
+		.replace("?size=4096", "?size=128")
+		.replace("?size=2048", "?size=128")
+		.replace("?size=1024", "?size=128")
+		.replace("?size=512", "?size=128")
+		.replace("?size=256", "?size=128")
+		.replace("?size=128", "?size=128")
+		.replace("?size=64", "?size=128");
+
+	if !url.ends_with("?size=128") {
+		url = format!("{}?size=128", url)
+	}
+
+	url
+}
 
 pub async fn calculate_users_color(
 	members: Vec<Member>, user_blacklist_server_image: Arc<RwLock<Vec<String>>>,
