+++ bot/src/background_task/server_image/generate_server_image.rs
@@ -16,9 +16,7 @@ use tokio::task;
 use tracing::{info, warn};
 use uuid::Uuid;
 
-use crate::background_task::server_image::calculate_user_color::{
-	get_image_from_url, get_member, return_average_user_color,
-};
+use crate::background_task::server_image::calculate_user_color::{change_to_x128_url, get_image_from_url, get_member, return_average_user_color};
 use crate::background_task::server_image::common::{
 	create_color_vector_from_tuple, create_color_vector_from_user_color, find_closest_color, Color,
 	ColorWithUrl,
@@ -28,7 +26,6 @@ use crate::constant::THREAD_POOL_SIZE;
 use crate::database::prelude::{ServerImage, UserColor};
 use crate::database::server_image::{ActiveModel, Column};
 use crate::helper::image_saver::general_image_saver::image_saver;
-use crate::new_member::change_to_x128_url;
 
 pub async fn generate_local_server_image(
 	ctx: &SerenityContext, guild_id: GuildId, image_config: ImageConfig,
