+++ bot/src/command/admin/server/mod.rs
@@ -1,3 +1,2 @@
 pub mod lang;
 pub mod module;
-pub mod new_member_setting;
